-Fixed segfault when an empty ccanvas was clicked (noted by Bas) -Removed some debug output on click.
git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@3164 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ tools/ltsview/lts.cpp
@@ -201,9 +201,9 @@ int LTS::getNumRanks() const {
 
 int LTS::getNumClusters() const {
   int result = 0;
-  for (unsigned int i = 0; i < clustersInRank.size(); ++i) {
+  for (size_t i = 0; i < clustersInRank.size(); ++i) {
     result += clustersInRank[i].size();
-	}
+  }
   return result;
 }
